Finish Blog Backend

+++ src/app.js
@@ -14,37 +14,43 @@ app.use(express.json());
 
 initdb();
 
-// app.post('/api/v1/users', validateUser, async (req, res) => {
-//     try{
-//         const {email, username, password} = req.body;
+app.post('/api/v1/users', validateUser, async (req, res) => {
+    try{
+        const {email, username, password} = req.body;
 
-//         const passwordHash = await bcrypt.hash(password, 10)
+        const passwordHash = await bcrypt.hash(password, 10)
 
-//         const insert = db.prepare(`INSERT INTO users(email, username, password_hash) VALUES (?, ?, ?)`);
+        const user = db.prepare(`
+            INSERT INTO users(email, username, password_hash) VALUES (?, ?, ?)
+        `).run(email, username, passwordHash);
 
-//         const info = insert.run(email, username, passwordHash)
+        return res.status(201).send({message: 'User succesfully created (201)'});
+    } catch (error) {
+        return res.status(400).send({message: `${error}`})
+    }
+})
 
-//         return res.status(201).send({message: `${info.lastInsertRowid}`});
-//     } catch (error) {
-//         console.log(error);
-//         return res.status(400).send({message: `${error}`})
-//     }
-// })
+app.post('/api/v1/auth/login', authenticateUser, async (req, res) => {
+    const payload = {
+        userID: req.userID
+    }
 
-// app.use('/api', authenticateCookie);
+    const token = jwt.sign(payload, process.env.SECRET_KEY, {
+        expiresIn: 60 * 60
+    });
 
-// app.post('/api/v1/auth/login', authenticateUser, async (req, res) => {
-//     const token = jwt.sign(req.body.username, process.env.SECRET_KEY);
-//     res.cookie('token', token, {
-//         httpOnly: true, 
-//         secure: true,
-//         sameSite: 'Lax',
-//         path: '/'
-//     });
+    res.cookie('token', token, {
+        httpOnly: true,
+        secure: false, 
+        sameSite: 'lax',
+        maxAge: 60 * 60 * 1000,
+        path: '/'
+    });
 
-//     res.status(201).send({message: `Succesfully logged in`})
-    
-// })
+    res.status(200).send({message: 'Login Successful'})
+})
+
+app.use(authenticateCookie);
 
 app.get('/api/v1/posts', (req, res) => {
     try{
@@ -83,14 +89,14 @@ app.get('/api/v1/posts/:id', (req, res) => {
 });
 
 app.post('/api/v1/posts', (req, res) => {
-    const {title, body, userID} = req.body; 
+    const {title, body} = req.body; 
+    const userID = req.userID; 
 
     try{
         const blog = db.prepare(`
             INSERT INTO blogs(title, body, userID) VALUES
 	            (?, ?, ?);    
         `).run(title, body, userID);
-
         return res.status(201).send({message: `Blog Created: ${blog.changes}`});
     } catch (error) {
         return res.status(400).send({error});
@@ -102,13 +108,22 @@ app.put('/api/v1/posts/:id', (req, res) => {
 
     try{
         const blog = db.prepare(`
+            SELECT B.title, B.body, B.publishDate, B.userID
+            FROM blogs B WHERE blogID = ?;
+        `).get(req.params.id);
+
+        if(blog.userID !== req.userID){
+            return res.status(401).send({message: 'Unauthorized request!'});
+        }
+
+        const newBlog = db.prepare(`
             UPDATE blogs
             SET
                 title = ?,
                 body = ?
             WHERE blogID = ?    
         `).run(title, body, req.params.id);
-        return res.status(200).send({message: `Blog Updated: ${blog.changes}`});
+        return res.status(200).send({message: `Blog Updated: ${newBlog.changes}`});
 
     } catch (error) {
         return res.status(400).send({error});
@@ -118,10 +133,19 @@ app.put('/api/v1/posts/:id', (req, res) => {
 app.delete('/api/v1/posts/:id', (req, res) => {
     try{
         const blog = db.prepare(`
+            SELECT B.title, B.body, B.publishDate, B.userID
+            FROM blogs B WHERE blogID = ?;
+        `).get(req.params.id);
+
+        if(blog.userID !== req.userID){
+            return res.status(401).send({message: 'Unauthorized request!'});
+        }
+
+        const deleteBlog = db.prepare(`
             DELETE FROM blogs
             WHERE blogID = ?
         `).run(req.params.id);
-        return res.status(204).send({message: 'blog deleted'});
+        return res.status(204).send({message: 'Blog deleted'});
     } catch(error) {
         return res.status(400).send({error});
     }
